+++ lib/theme/theme.dart
@@ -1,0 +1,5 @@
+library theme;
+
+export 'app_theme.dart';
+export 'background.dart';
+export 'gradient_text.dart';
